splitted component & model & removed state from host
