+++ yew-ui/src/components/host/msg.rs
@@ -1,0 +1,3 @@
+pub enum Msg {
+    Start,
+}
